tests/submission/PhilipTaylor/canvas/2d.line.cap.square.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.line.cap.square</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.<a href="index.2d.line.cap.html">cap</a>.square</h1>
<p class="desc">lineCap 'square' is rendered correctly</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.lineCap.square">2d.lineCap.square</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("lineCap 'square' is rendered correctly");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

ctx.lineCap = 'square';
ctx.lineWidth = 20;

ctx.fillStyle = '#f00';
ctx.strokeStyle = '#0f0';
ctx.fillRect(15, 5, 20, 40);
ctx.beginPath();
ctx.moveTo(25, 15);
ctx.lineTo(25, 35);
ctx.stroke();

ctx.fillStyle = '#0f0';
ctx.strokeStyle = '#f00';
ctx.beginPath();
ctx.moveTo(75, 15);
ctx.lineTo(75, 35);
ctx.stroke();
ctx.fillRect(65, 5, 20, 40);

_assertPixel(canvas, 25,4, 0,255,0,255, "25,4", "0,255,0,255");
_assertPixel(canvas, 25,5, 0,255,0,255, "25,5", "0,255,0,255");
_assertPixel(canvas, 25,6, 0,255,0,255, "25,6", "0,255,0,255");
_assertPixel(canvas, 25,44, 0,255,0,255, "25,44", "0,255,0,255");
_assertPixel(canvas, 25,45, 0,255,0,255, "25,45", "0,255,0,255");
_assertPixel(canvas, 25,46, 0,255,0,255, "25,46", "0,255,0,255");

_assertPixel(canvas, 75,4, 0,255,0,255, "75,4", "0,255,0,255");
_assertPixel(canvas, 75,5, 0,255,0,255, "75,5", "0,255,0,255");
_assertPixel(canvas, 75,6, 0,255,0,255, "75,6", "0,255,0,255");
_assertPixel(canvas, 75,44, 0,255,0,255, "75,44", "0,255,0,255");
_assertPixel(canvas, 75,45, 0,255,0,255, "75,45", "0,255,0,255");
_assertPixel(canvas, 75,46, 0,255,0,255, "75,46", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
